+++ dialogue2graph/metrics/llm_validators/validators.py
@@ -6,7 +6,6 @@ This module contains validators to evaluate dialogs
 """
 
 from typing import List
-import re
 
 from pydantic import BaseModel, Field
 
@@ -55,10 +54,6 @@ START_THRESHOLD = 0.2
 END_THRESHOLD = 0.2
 
 
-def _message_has_greeting_re(regex: str, text: str) -> bool:
-    return bool(re.match(regex, text, flags=re.IGNORECASE))
-
-
 def _message_has_greeting_llm(model: BaseChatModel, text: str) -> bool:
 
     class OpeningValidation(BaseModel):
@@ -82,10 +77,6 @@ def _message_has_greeting_llm(model: BaseChatModel, text: str) -> bool:
     return result.isOpening
 
 
-def _message_has_closing_re(regex: str, text: str) -> bool:
-    return bool(re.search(regex, text, flags=re.IGNORECASE))
-
-
 def _message_has_closing_llm(model: BaseChatModel, text: str) -> bool:
 
     class ClosingValidation(BaseModel):
@@ -107,43 +98,6 @@ def _message_has_closing_llm(model: BaseChatModel, text: str) -> bool:
     closing_val_chain = close_prompt | model | parser
     result = closing_val_chain.invoke({"text": text})
     return result.isClosing
-
-
-def is_greeting_repeated_regex(dialogs: List[Dialogue], regex: str = None) -> bool:
-    """
-    Checks if greeting is repeated within dialogues using regular expression.
-    Args:
-        dialogs (List[Dialogue]): Dialog list from graph.
-        regex (str): Regular expression to find start turns. Defaults to None, so standard regex is used.
-    Returns
-        bool: True if greeting has been repeated, False otherwise.
-    """
-    if not regex:
-        regex = r"^hello|^hi|^greetings"
-    for dialog in dialogs:
-        for i, message in enumerate(dialog.messages):
-            if i != 0 and message.participant == "assistant" and _message_has_greeting_re(regex, message.text):
-                return True
-    return False
-
-
-def is_dialog_closed_too_early_regex(dialogs: List[Dialogue], regex: str = None) -> bool:
-    """
-    Checks if assistant tried to close dialogue in the middle using regular expression.
-    Args:
-        dialogs (List[Dialogue]): Dialog list from graph.
-        regex (str): Regular expression to find end turns. Defaults to None, so standard regex is used.
-    Returns
-        bool: True if closing appeared too early, False otherwise.
-    """
-    if not regex:
-        regex = r"have a (great|good|nice) day.$|goodbye.$"
-    for dialog in dialogs:
-        last_turn_idx = len(dialog.messages) - 1
-        for i, message in enumerate(dialog.messages):
-            if i != last_turn_idx and message.participant == "assistant" and _message_has_closing_re(regex, message.text):
-                return True
-    return False
 
 
 def is_greeting_repeated_emb_llm(
